Extract aggregate root into traits
