+++ src/Aggregate/EventProducerTrait.php
@@ -1,0 +1,63 @@
+<?php
+/**
+ * This file is part of the prooph/event-sourcing.
+ * (c) 2014-2017 prooph software GmbH <contact@prooph.de>
+ * (c) 2015-2017 Sascha-Oliver Prolic <saschaprolic@googlemail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Prooph\EventSourcing\Aggregate;
+
+use Prooph\EventSourcing\AggregateChanged;
+
+trait EventProducerTrait
+{
+    /**
+     * Current version
+     *
+     * @var int
+     */
+    protected $version = 0;
+
+    /**
+     * List of events that are not committed to the EventStore
+     *
+     * @var AggregateChanged[]
+     */
+    protected $recordedEvents = [];
+
+    /**
+     * Get pending events and reset stack
+     *
+     * @return AggregateChanged[]
+     */
+    protected function popRecordedEvents(): array
+    {
+        $pendingEvents = $this->recordedEvents;
+
+        $this->recordedEvents = [];
+
+        return $pendingEvents;
+    }
+
+    /**
+     * Record an aggregate changed event
+     */
+    protected function recordThat(AggregateChanged $event): void
+    {
+        $this->version += 1;
+
+        $this->recordedEvents[] = $event->withVersion($this->version);
+
+        $this->apply($event);
+    }
+
+    /**
+     * Apply given event
+     */
+    abstract protected function apply(AggregateChanged $event): void;
+}
